Fixes #80 -- Use render_variations field kwarg in management command
Previously the management command ignored the `render_variations`
kwarg on the `StdImageField`. This was a problem if someone used
a custom callable that did the rendering manually.

The management command now obeys the `render_variations`. Should it
be a callable, the callable will be called for all images in the
separate subprocesses.

Closes #81