Set file stream curstor to zero position after running validators (#218)
The `BaseSizeValidator` read the file stream to create an in memory Pillow Image object. In reading the file, the file streams cursor was set to the end. Therefore, the validator left the file in a different state than receiving it. This patch sets the cursor back to zero after reading the file.